backend/requirements.txt
fastapi==0.115.6
uvicorn==0.22.0
python-dotenv==1.0.1
pydantic==2.10.4
sqlalchemy==2.0.36
alembic==1.10.0
pytest==8.3.4
httpx==0.28.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.20
transformers==4.47.1
torch==2.5.1
sentencepiece==0.1.97
datasets==2.12.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
whisper==1.0.0
pytube==15.0.0
psycopg2-binary==2.9.6
pydub==0.25.1
openai==0.27.8
tiktoken==0.4.0
tqdm==4.65.0
openai-whisper @ git+https://github.com/openai/whisper.git@main
